smart_buffet_planner/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.5.1
django-cors-headers==4.3.1
Pillow==10.1.0
reportlab==4.0.7
python-dateutil==2.8.2
django-filter==23.3
whitenoise==6.6.0
setuptools
